Fix /apis

+++ frontend/src/javascripts/shared/apis/GoodApi.ts
@@ -1,12 +1,15 @@
 import axios from './common/axios';
 
-export const addGoodList = (id, type) => {
+export const addGoodList = (id: number, type: string) => {
+  const formatPoatType = type.charAt(0).toUpperCase() + type.slice(1);
   axios.post('/goods', {
-    id,
-    type
+    post_id: id,
+    post_type: formatPoatType
   });
 };
 
-export const removeGoodList = id => {
-  axios.delete(`/goods/${id}`);
+export const removeGoodList = (id: number | null) => {
+  if (typeof id == 'number') {
+    axios.delete(`/goods/${id}`);
+  }
 };
